+++ src/main/java/math/Vector3f.kt
@@ -6,15 +6,15 @@ import kotlin.math.sin
 import kotlin.math.sqrt
 
 class Vector3f(
-    private var x: Double,
-    private var y: Double,
-    private var z:Double
+    private var x: Float,
+    private var y: Float,
+    private var z:Float
 ) : MatrixBuilder(arrayOf(arrayOf(x, y,z))) {
 
     val length = 3
 
     // auxiliary constructor
-    constructor(value: Double) : this(value, value,value)
+    constructor(value: Float) : this(value, value,value)
 
     fun normalize() {
         val length = sqrt(x.pow(2) + y.pow(2) + z.pow(2))
@@ -23,25 +23,25 @@ class Vector3f(
         z /= length
     }
 
-    fun rotate(theta: Double) {
+    fun rotate(theta: Float) {
         x = x * cos(theta) - y * sin(theta)
         y = x * sin(theta) + y * cos(theta)
     }
 
     // abstract members' implementation
-    override fun plusAssign(scalar: Double) {
+    override fun plusAssign(scalar: Float) {
         this += Vector3f(scalar)
     }
 
-    override fun minusAssign(scalar: Double) {
+    override fun minusAssign(scalar: Float) {
         this -= Vector3f(scalar)
     }
 
-    override fun plus(scalar: Double) = (this + Vector3f(scalar))!!
+    override fun plus(scalar: Float) = (this + Vector3f(scalar))!!
 
-    override fun minus(scalar: Double) = (this - Vector3f(scalar))!!
+    override fun minus(scalar: Float) = (this - Vector3f(scalar))!!
 
-    override fun inc() = this + 1.0
+    override fun inc() = this + 1.0f
 
-    override fun dec() = this - 1.0
+    override fun dec() = this - 1.0f
 }
